fixing bugs, I promise the decomps will be perfect

+++ RSDKv3/Drawing.cpp
@@ -6813,8 +6813,8 @@ void DrawTextMenuEntry(void *menu, int rowID, int XPos, int YPos, int textHighli
     TextMenu *tMenu = (TextMenu *)menu;
     int id          = tMenu->entryStart[rowID];
     for (int i = 0; i < tMenu->entrySize[rowID]; ++i) {
-        DrawSprite(XPos + (i << 3), YPos, 8, 8, (int)((int)(tMenu->textData[id] & 0xF) << 3),
-                   (int)((int)(tMenu->textData[id] >> 4) << 3) + textHighlight, textMenuSurfaceNo);
+        DrawSprite(XPos + (i << 3), YPos, 8, 8, ((tMenu->textData[id] & 0xF) << 3), ((tMenu->textData[id] >> 4) << 3) + textHighlight,
+                   textMenuSurfaceNo);
         id++;
     }
 }
@@ -6824,12 +6824,11 @@ void DrawStageTextEntry(void *menu, int rowID, int XPos, int YPos, int textHighl
     int id          = tMenu->entryStart[rowID];
     for (int i = 0; i < tMenu->entrySize[rowID]; ++i) {
         if (i == tMenu->entrySize[rowID] - 1) {
-            DrawSprite(XPos + (i << 3), YPos, 8, 8, (int)((int)(tMenu->textData[id] & 0xF) << 3), (int)((int)(tMenu->textData[id] >> 4) << 3),
-                       textMenuSurfaceNo);
+            DrawSprite(XPos + (i << 3), YPos, 8, 8, ((tMenu->textData[id] & 0xF) << 3), ((tMenu->textData[id] >> 4) << 3), textMenuSurfaceNo);
         }
         else {
-            DrawSprite(XPos + (i << 3), YPos, 8, 8, (int)((int)(tMenu->textData[id] & 0xF) << 3),
-                       (int)((int)(tMenu->textData[id] >> 4) << 3) + textHighlight, textMenuSurfaceNo);
+            DrawSprite(XPos + (i << 3), YPos, 8, 8, ((tMenu->textData[id] & 0xF) << 3), ((tMenu->textData[id] >> 4) << 3) + textHighlight,
+                       textMenuSurfaceNo);
         }
         id++;
     }
@@ -6839,8 +6838,8 @@ void DrawBlendedTextMenuEntry(void *menu, int rowID, int XPos, int YPos, int tex
     TextMenu *tMenu = (TextMenu *)menu;
     int id          = tMenu->entryStart[rowID];
     for (int i = 0; i < tMenu->entrySize[rowID]; ++i) {
-        DrawBlendedSprite(XPos + (i << 3), YPos, 8, 8, (int)((int)(tMenu->textData[id] & 0xF) << 3),
-                          (int)((int)(tMenu->textData[id] >> 4) << 3) + textHighlight, textMenuSurfaceNo);
+        DrawBlendedSprite(XPos + (i << 3), YPos, 8, 8, ((tMenu->textData[id] & 0xF) << 3), ((tMenu->textData[id] >> 4) << 3) + textHighlight,
+                          textMenuSurfaceNo);
         id++;
     }
 }
@@ -6888,6 +6887,7 @@ void DrawTextMenu(void *menu, int XPos, int YPos)
                             DrawTextMenuEntry(tMenu, i, XPos, YPos, 128);
                         else
                             DrawTextMenuEntry(tMenu, i, XPos, YPos, 0);
+
                         if (i == tMenu->selection2 && i != tMenu->selection1)
                             DrawStageTextEntry(tMenu, i, XPos, YPos, 128);
                         break;
@@ -6906,17 +6906,20 @@ void DrawTextMenu(void *menu, int XPos, int YPos)
                         else
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 0);
                         break;
+
                     case 2:
                         if (i == tMenu->selection1 || i == tMenu->selection2)
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 128);
                         else
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 0);
                         break;
+
                     case 3:
                         if (i == tMenu->selection1)
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 128);
                         else
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 0);
+
                         if (i == tMenu->selection2 && i != tMenu->selection1)
                             DrawStageTextEntry(tMenu, i, entryX, YPos, 128);
                         break;
@@ -6935,12 +6938,14 @@ void DrawTextMenu(void *menu, int XPos, int YPos)
                         else
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 0);
                         break;
+
                     case 2:
                         if (i == tMenu->selection1 || i == tMenu->selection2)
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 128);
                         else
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 0);
                         break;
+
                     case 3:
                         if (i == tMenu->selection1)
                             DrawTextMenuEntry(tMenu, i, entryX, YPos, 128);
